+++ app/src/main/java/com/maxor/raveltie/score/ScoreResponse.kt
@@ -1,4 +1,4 @@
 package com.maxor.raveltie.score
 
-data class ScoreResponse ( val score: Int){
+data class ScoreResponse ( val score: Float){
 }
